fixing some warnings

+++ tests/test.c
@@ -138,7 +138,7 @@ int main(int argc, const char* argv[]) {
     int last_x = 0, last_y = 240;
     for (long i = sine_i; i < (sine_i + 641); ++i) {
       float x = (float)(i - sine_i);
-      float y = 240.f + (100.f * sin(i * (3.141 / 180.f)));
+      float y = 240.f + (100.f * sinf(i * (3.141f / 180.f)));
       line(win, last_x, last_y, x, y, col);
       last_x = x;
       last_y = y;
